gid is now stable and unique
This changes is required, to make the id stable and unique for each product. Any history you might have in your merchant center feed or anywhere else might be lost, since the ID in the feed for each product will change!

This change is required to make it possible to use the feed for remarketing.

https://support.google.com/merchants/answer/6324405